completed user register api

+++ src/main/java/com/esun/library/domain/entity/User.java
@@ -2,11 +2,8 @@ package com.esun.library.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 import java.util.Date;
-import java.util.List;
 
 @Entity
 @Data
@@ -14,7 +11,7 @@ import java.util.List;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;
+    private Integer userId;
 
     @Column(unique = true, nullable = false)
     private String phoneNumber;
